Refactored gitlib and commit tracking
Created GitLib library to encapsulate code for managing git objects.  All git-specific classes moved there.  A unit test project was created to test this.

All Vault commits and Git commits are loaded and current progress is saved to an XML file.  Resuming a conversion now loads exclusively from the XML file and conversion data is no longer saved to the git commit messages.

Merge commits are processed properly as part of the process, so grafts are no longer neccessary.  Code for creating the grafts was removed.