Added ability to process multiple repositories
You can now configure a repository XML file with specific details for each repo in each line.  The processor will process each repo in a sequence.

The commit also split the branch renaming file and the author map file.  The author map is the same in every run, while the branch renaming may be different for each repo.  If updates are required for author mapping, it only needs to be done in the one location instead of in the mapping file for each repo.